+++ lib/Model/UpdateDealRequest.php
@@ -73,6 +73,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         'pipeline_id' => 'int',
         'stage_id' => 'int',
         'status' => 'string',
+        'won_time' => 'string',
+        'lost_time' => 'string',
+        'close_time' => 'string',
         'expected_close_date' => '\DateTime',
         'probability' => 'float',
         'lost_reason' => 'string',
@@ -97,6 +100,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         'pipeline_id' => null,
         'stage_id' => null,
         'status' => null,
+        'won_time' => null,
+        'lost_time' => null,
+        'close_time' => null,
         'expected_close_date' => 'date',
         'probability' => null,
         'lost_reason' => null,
@@ -144,6 +150,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         'pipeline_id' => 'pipeline_id',
         'stage_id' => 'stage_id',
         'status' => 'status',
+        'won_time' => 'won_time',
+        'lost_time' => 'lost_time',
+        'close_time' => 'close_time',
         'expected_close_date' => 'expected_close_date',
         'probability' => 'probability',
         'lost_reason' => 'lost_reason',
@@ -166,6 +175,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         'pipeline_id' => 'setPipelineId',
         'stage_id' => 'setStageId',
         'status' => 'setStatus',
+        'won_time' => 'setWonTime',
+        'lost_time' => 'setLostTime',
+        'close_time' => 'setCloseTime',
         'expected_close_date' => 'setExpectedCloseDate',
         'probability' => 'setProbability',
         'lost_reason' => 'setLostReason',
@@ -188,6 +200,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         'pipeline_id' => 'getPipelineId',
         'stage_id' => 'getStageId',
         'status' => 'getStatus',
+        'won_time' => 'getWonTime',
+        'lost_time' => 'getLostTime',
+        'close_time' => 'getCloseTime',
         'expected_close_date' => 'getExpectedCloseDate',
         'probability' => 'getProbability',
         'lost_reason' => 'getLostReason',
@@ -292,6 +307,9 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
         $this->container['pipeline_id'] = $data['pipeline_id'] ?? null;
         $this->container['stage_id'] = $data['stage_id'] ?? null;
         $this->container['status'] = $data['status'] ?? null;
+        $this->container['won_time'] = $data['won_time'] ?? null;
+        $this->container['lost_time'] = $data['lost_time'] ?? null;
+        $this->container['close_time'] = $data['close_time'] ?? null;
         $this->container['expected_close_date'] = $data['expected_close_date'] ?? null;
         $this->container['probability'] = $data['probability'] ?? null;
         $this->container['lost_reason'] = $data['lost_reason'] ?? null;
@@ -579,6 +597,78 @@ class UpdateDealRequest implements ModelInterface, ArrayAccess, JsonSerializable
             );
         }
         $this->container['status'] = $status;
+
+        return $this;
+    }
+
+    /**
+     * Gets won_time
+     *
+     * @return string|null
+     */
+    public function getWonTime()
+    {
+        return $this->container['won_time'];
+    }
+
+    /**
+     * Sets won_time
+     *
+     * @param string|null $won_time The optional date and time of changing the deal status as won in UTC. Format: YYYY-MM-DD HH:MM:SS. Can be set only when deal `status` is already Won. Can not be used together with `lost_time`.
+     *
+     * @return self
+     */
+    public function setWonTime($won_time): self
+    {
+        $this->container['won_time'] = $won_time;
+
+        return $this;
+    }
+
+    /**
+     * Gets lost_time
+     *
+     * @return string|null
+     */
+    public function getLostTime()
+    {
+        return $this->container['lost_time'];
+    }
+
+    /**
+     * Sets lost_time
+     *
+     * @param string|null $lost_time The optional date and time of changing the deal status as lost in UTC. Format: YYYY-MM-DD HH:MM:SS. Can be set only when deal `status` is already Lost. Can not be used together with `won_time`.
+     *
+     * @return self
+     */
+    public function setLostTime($lost_time): self
+    {
+        $this->container['lost_time'] = $lost_time;
+
+        return $this;
+    }
+
+    /**
+     * Gets close_time
+     *
+     * @return string|null
+     */
+    public function getCloseTime()
+    {
+        return $this->container['close_time'];
+    }
+
+    /**
+     * Sets close_time
+     *
+     * @param string|null $close_time The optional date and time of closing the deal in UTC. Format: YYYY-MM-DD HH:MM:SS.
+     *
+     * @return self
+     */
+    public function setCloseTime($close_time): self
+    {
+        $this->container['close_time'] = $close_time;
 
         return $this;
     }